pyproject.toml
[project]
name = "taskflow-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "openai>=2.31.0",
    "pinecone>=8.1.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
